requirements.txt
atomicwrites==1.4.0
attrs==21.4.0
colorama==0.4.4
iniconfig==1.1.1
packaging==21.3
pluggy==1.0.0
py==1.11.0
pyparsing==3.0.7
pytest==7.0.0
tomli==2.0.0
selenium==4.1.0
webdriver-manager==3.5.4
names==0.3.0
pytest-xdist==2.5.0
allure-pytest
